requirements_frozen.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
catboost==1.2.8
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
coverage==7.13.3
cryptography==46.0.4
cycler==0.12.1
defusedxml==0.7.1
Deprecated==1.3.1
docx2pdf==0.1.8
easyocr==1.7.2
ecdsa==0.19.1
et_xmlfile==2.0.0
fastapi==0.128.2
filelock==3.20.3
flatbuffers==25.12.19
fonttools==4.61.1
fpdf2==2.8.5
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
graphviz==0.21
greenlet==3.3.1
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
ImageIO==2.37.2
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lazy_loader==0.4
ldap3==2.9.1
limits==5.8.0
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
ml_dtypes==0.5.4
mpmath==1.3.0
narwhals==2.16.0
networkx==3.6.1
ninja==1.13.0
numpy==2.4.2
onnx==1.20.1
onnxruntime==1.24.1
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
optimum==2.1.0
optimum-onnx==0.1.0
packaging==26.0
pandas==3.0.0
passlib==1.7.4
pdf2image==1.17.0
pillow==12.1.0
plotly==6.5.2
pluggy==1.6.0
proto-plus==1.27.1
protobuf==5.29.6
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyMuPDF==1.26.7
pyparsing==3.3.2
pypdf==6.6.2
pytest==9.0.2
pytest-cov==7.0.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
python-pptx==1.0.2
pywin32==311
PyYAML==6.0.3
RapidFuzz==3.14.3
redis==7.2.1
regex==2026.1.15
requests==2.32.5
rsa==4.9.1
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
setuptools==80.10.2
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
SQLAlchemy==2.0.46
starlette==0.50.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2026.1.28
tokenizers==0.22.2
torch==2.10.0+cpu
torchvision==0.25.0+cpu
tqdm==4.67.3
transformers==4.57.6
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0
wrapt==2.1.1
xlrd==2.0.2
xlsxwriter==3.2.9
